Changed the format of the LoginTestCase objects. This makes it easier to define the test case and allows much more flexibility in what you can test. LoginAttempt classes no longer need to be implemented.
Also moved the SAML2Int's mock IdP to the abstract SPTestSuite so every
test suite already has a mock IdP to work with. You only have to change
it if you have specific needs for it. This should make it easier to
start writing test cases in a test suite.

The readme was updated to reflect these changes