Updated the NotOrOrAfter test case.
It now also checks if long validity periods can also be used. It's
possible for an SP to use their own max validity period that sometimes
overrides the NotOnOrAfter time of the SAML message. This could shorten
the actual period that the message is valid and contradicts the value
given in the NotOnOrAfter attribute.